Paging

+++ src/main/java/myBlog/api/response/PostResponse.java
@@ -4,12 +4,11 @@ import java.util.List;
 import lombok.Data;
 
 import myBlog.dto.PostDto;
-import myBlog.model.Post;
 
 @Data
 public class PostResponse {
 
-    private int count;
+    private Long count;
 
     private List<PostDto> posts;
